pyproject.toml
[project]
name = "mcp-server-iris"
version = "0.3.1"
description = "A Model Context Protocol server for InterSystems IRIS."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "CaretDev Corp." }]
maintainers = [{ name = "Dmitry Maslennikov", email = "dmitry@caretdev.com" }]
keywords = ["iris", "mcp", "llm", "automation"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "intersystems-irispython>=5.1.0",
    "mcp[cli]>=1.2.0",
    "starlette>=0.36.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
mcp-server-iris = "mcp_server_iris:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "ruff>=0.7.3", "pytest>=8.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
